# openapi.yaml
openapi: 3.0.3
info:
  title: SafeGirl Backend API
  description: |
    Women's safety reporting platform with encrypted reports, blockchain verification,
    non-blocking panic alerts, and report sharing with access control.

    **Key Features:**
    - Phone-only OTP authentication (no passwords)
    - AES-256-GCM encryption for report data
    - IPFS + Polygon blockchain for immutable report records
    - Non-blocking panic alerts with SMS notifications
    - Report access control and sharing
    - Key recovery with PIN backup
  version: 1.0.0
  contact:
    name: SafeGirl Team
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.safegirl.example.com
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from signup/login verify endpoints
  schemas:
...
    SubmitReportRequest:
      type: object
      required: [payload]
      properties:
        payload:
          type: object
          description: Unencrypted report data (backend encrypts before IPFS upload)
          example:
            title: "Safety Incident"
            description: "Incident details"
            location: "Downtown"
            timestamp: 1234567890
        responses:
          type: array
          items:
            type: string
          description: Responses to survey questions (optional)
          example: ["Yes", "It happened at work", "Yesterday", "For personal tracking", "No"]
        metadata:
          type: object
          description: Optional metadata
          example:
            category: "harassment"
            severity: "medium"
    SubmitReportResponse:
      type: object
      required: [success, reportId, status]
      properties:
        success:
          type: boolean
          example: true
        reportId:
          type: string
          description: Unique report identifier
          example: "550e8400-e29b-41d4-a716-446655440000"
        status:
          type: string
          enum: [pending, confirmed, failed]
          example: "pending"
        ipfsHash:
          type: string
          nullable: true
          description: IPFS hash once confirmed
        txHash:
          type: string
          nullable: true
          description: Blockchain transaction hash once confirmed
    ReportStatusResponse:
      type: object
      required: [reportId, status]
      properties:
        reportId:
          type: string
        status:
          type: string
          enum: [pending, confirmed, failed]
          example: "confirmed"
        ipfsHash:
          type: string
          nullable: true
        txHash:
          type: string
          nullable: true
        blockNumber:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        confirmedAt:
          type: string
          format: date-time
          nullable: true
    DecryptedReportResponse:
      type: object
      required: [reportId, payload]
      properties:
        reportId:
          type: string
        payload:
          type: object
          description: Decrypted report data
        responses:
          type: array
          items:
            type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /api/submitReport:
    post:
      summary: Submit encrypted report
      operationId: submitReport
      tags: [Reports]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitReportRequest'
      responses:
        '200':
          description: Report submitted (pending blockchain confirmation)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmitReportResponse'
        '400':
          description: Invalid payload
        '401':
          description: Unauthorized
        '503':
          description: IPFS or blockchain service unavailable
  /api/reportStatus:
    get:
      summary: Check report status (pending/confirmed/failed)
      operationId: getReportStatus
      tags: [Reports]
      security:
        - bearerAuth: []
      parameters:
        - name: reportId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Report status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportStatusResponse'
        '404':
          description: Report not found
  /api/report/{reportId}/decrypt:
    get:
      summary: Decrypt and view report (owner or granted access)
      operationId: getDecryptedReport
      tags: [Reports]
      security:
        - bearerAuth: []
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Decrypted report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecryptedReportResponse'
        '403':
          description: Access denied
        '404':
          description: Report not found
  /api/report/{reportId}/archive:
    post:
      summary: Archive report (hide from library, keep on blockchain)
      operationId: archiveReport
      tags: [Reports]
      security:
        - bearerAuth: []
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Report archived
        '404':
          description: Report not found
  /api/report/{reportId}/unarchive:
    post:
      summary: Unarchive report
      operationId: unarchiveReport
      tags: [Reports]
      security:
        - bearerAuth: []
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Report unarchived
...